test(core): reorganize unit tests by module under test
Dissolve the streaming-regressions and async-writes grab-bags into the
files of the modules they pin, split test_dataset_streaming.py into its
dataset / data_manager / read-dispatcher parts, and share the streaming
stub and TTA driver through tests/unit/conftest.py.
